[API] Rename namespace "/gc" into "/system"
