+++ api/lib/application/system/index.js
@@ -1,4 +1,4 @@
-const gcController = require('./gc-controller');
+const systemController = require('./system-controller');
 const securityController = require('../../interfaces/controllers/security-controller');
 
 exports.register = async function(server) {
@@ -6,13 +6,13 @@ exports.register = async function(server) {
   server.route([
     {
       method: 'GET',
-      path: '/api/gc/heap-dump',
+      path: '/api/system/heap-dump',
       config: {
         pre: [{
           method: securityController.checkUserHasRolePixMaster,
           assign: 'hasRolePixMaster'
         }],
-        handler: gcController.generateAndDownloadHeapDump,
+        handler: systemController.generateAndDownloadHeapDump,
         notes: [
           '- **Route nécessitant une authentification en tant que Pix Master**\n' +
           '- Génère et retourne un fichier heap dump'
@@ -22,13 +22,13 @@ exports.register = async function(server) {
     },
     {
       method: 'GET',
-      path: '/api/gc/heap-profile',
+      path: '/api/system/heap-profile',
       config: {
         pre: [{
           method: securityController.checkUserHasRolePixMaster,
           assign: 'hasRolePixMaster'
         }],
-        handler: gcController.generateAndDownloadHeapProfile,
+        handler: systemController.generateAndDownloadHeapProfile,
         notes: [
           '- **Route nécessitant une authentification en tant que Pix Master**\n' +
           '- Génère et retourne un fichier heap profile'
