+++ api/tests/integration/application/system/system-controller_test.js
@@ -1,15 +1,15 @@
 const { expect, sinon, HttpTestServer } = require('../../../test-helper');
 const securityController = require('../../../../lib/interfaces/controllers/security-controller');
-const gcController = require('../../../../lib/application/gc/gc-controller');
-const moduleUnderTest = require('../../../../lib/application/gc');
+const systemController = require('../../../../lib/application/system/system-controller');
+const moduleUnderTest = require('../../../../lib/application/system');
 
-describe('Integration | Application | GC | gc-controller', () => {
+describe('Integration | Application | System | system-controller', () => {
 
   let httpTestServer;
 
   beforeEach(() => {
-    sinon.stub(gcController, 'generateAndDownloadHeapDump').resolves('any Node.js Stream');
-    sinon.stub(gcController, 'generateAndDownloadHeapProfile').resolves('any Node.js Stream');
+    sinon.stub(systemController, 'generateAndDownloadHeapDump').resolves('any Node.js Stream');
+    sinon.stub(systemController, 'generateAndDownloadHeapProfile').resolves('any Node.js Stream');
     sinon.stub(securityController, 'checkUserHasRolePixMaster');
     httpTestServer = new HttpTestServer(moduleUnderTest);
   });
@@ -28,7 +28,7 @@ describe('Integration | Application | GC | gc-controller', () => {
 
       it('should resolve a 200 HTTP response', async () => {
         // when
-        const response = await httpTestServer.request('GET', '/api/gc/heap-dump');
+        const response = await httpTestServer.request('GET', '/api/system/heap-dump');
 
         // then
         expect(response.statusCode).to.equal(200);
@@ -47,7 +47,7 @@ describe('Integration | Application | GC | gc-controller', () => {
 
         it('should resolve a 403 HTTP response', () => {
           // when
-          const promise = httpTestServer.request('GET', '/api/gc/heap-dump');
+          const promise = httpTestServer.request('GET', '/api/system/heap-dump');
 
           // then
           return promise.then((response) => {
@@ -68,7 +68,7 @@ describe('Integration | Application | GC | gc-controller', () => {
 
       it('should resolve a 200 HTTP response', async () => {
         // when
-        const response = await httpTestServer.request('GET', '/api/gc/heap-profile');
+        const response = await httpTestServer.request('GET', '/api/system/heap-profile');
 
         // then
         expect(response.statusCode).to.equal(200);
@@ -87,7 +87,7 @@ describe('Integration | Application | GC | gc-controller', () => {
 
         it('should resolve a 403 HTTP response', () => {
           // when
-          const promise = httpTestServer.request('GET', '/api/gc/heap-profile');
+          const promise = httpTestServer.request('GET', '/api/system/heap-profile');
 
           // then
           return promise.then((response) => {
